documentation and small changes to Gateway interfaces to make it easier for others to call getAccounts() while maintaining CA

+++ src/main/java/screens/login_registration/FileUser.java
@@ -36,10 +36,8 @@ public class FileUser implements UserRegGateway, LoginGateway, LogoutGateway, Co
         // and reads the existing file if it does exist
 
         if (Files.exists(Path.of(path))) {
-//            this.accounts = readFile();
             accounts = readFile();
         } else {
-//            this.accounts = new HashMap<String, UserRegSaveRequest>();
             accounts = new HashMap<String, UserRegSaveRequest>();
             save();
         }
@@ -69,13 +67,14 @@ public class FileUser implements UserRegGateway, LoginGateway, LogoutGateway, Co
          * Write the map of usernames to UserRegSaveRequest objects into the User database file.
          */
 
+        // check if file exists already and if it doesn't, open a new file
         FileOutputStream fileWriter;
-//        if (Files.exists(Path.of(filePath))) {
-//            fileWriter = new FileOutputStream(filePath);
-//        } else {
-//            fileWriter = new FileOutputStream("src/main/java/data/users.ser");
-//        }
-        fileWriter = new FileOutputStream(filePath);
+        if (Files.exists(Path.of(filePath))) {
+            fileWriter = new FileOutputStream(filePath);
+        } else {
+            fileWriter = new FileOutputStream("src/main/java/data/users.ser");
+        }
+//        fileWriter = new FileOutputStream(filePath);
         ObjectOutputStream out = new ObjectOutputStream(fileWriter);
         out.writeObject(accounts);
         out.close();
@@ -97,8 +96,8 @@ public class FileUser implements UserRegGateway, LoginGateway, LogoutGateway, Co
         return accounts.get(name).getPass();
     }
 
+    @Override
     public Map<String, UserRegSaveRequest> getAccounts() {
-//        return this.accounts;
         return accounts;
     }
